A shell function that creates a scratch repo.
git-scratch creates git-scratch.d/, for experiments,
with an initialized repo and one, committed, empty file.

On termination, you're pushd'd to that directory to start work.

Don't make it an executable -- it's a function, not a script.
Putting this in your path makes it sourceable.
(e.g. ln $PWD/git-scratch ~/bin; source git-scratch)